test(sellproduct): 精简生成器回归测试
